+++ src/utils/environment.ts
@@ -5,12 +5,10 @@ import { constructAPI, IHermesAPI } from "./Hermes";
 declare var Drupal: any;
 
 interface IUrls {
-  hermesBaseUrl: string;
+  hermesApiUrl: string;
   orderNotificationHost: string;
   orderNotificationPath: string;
-  hermesOrderUrl: string;
   profileUrl: string;
-  hermesScriptUrl: string;
 }
 
 export interface IEnvironment {
@@ -34,13 +32,13 @@ function getEnvironmentDependentURLs() {
   if (getEnvironment() === "dev") {
     const devPostfix: string = window.location.hostname.split(".").slice(-5).join(".");
     return {
-      hermesBaseUrl: `${window.location.hostname}`.replace("nsidc.org.drupal", "hermes2"),
+      hermesApiUrl: `/order-proxy/api-v5`,
       orderNotificationHost: `wss://dev.hermes2.${devPostfix}`,
       orderNotificationPath: "/notification/",
     };
   } else {
     return {
-      hermesBaseUrl: `${window.location.hostname}/apps/orders`,
+      hermesApiUrl: `/order-proxy/api-v5`,
       orderNotificationHost: `wss://${window.location.hostname}`,
       orderNotificationPath: "/apps/orders/notification2/",
     };
@@ -71,31 +69,26 @@ export default function setupEnvironment(inDrupal: boolean): IEnvironment {
   if (inDrupal) {
     const urls = {
       ...getEnvironmentDependentURLs(),
-      hermesOrderUrl: "/order-proxy",
-      hermesScriptUrl: "/script-proxy",
       profileUrl: "/order-history",
     };
     return {
       drupalDataset: Drupal.settings.data_downloads.dataset,
       exposeFunction,
-      hermesAPI: constructAPI(urls, true),
+      hermesAPI: constructAPI(urls),
       inDrupal,
       urls,
       user: Drupal.settings.data_downloads.user,  // TODO: Use the Earthdata Login module function?
     };
   } else {
     const environmentDependentURLs = getEnvironmentDependentURLs();
-    const hermesBaseUrl = environmentDependentURLs.hermesBaseUrl;
     const urls = {
       ...environmentDependentURLs,
-      hermesOrderUrl: `https://${hermesBaseUrl}/api-v5/orders/`,
-      hermesScriptUrl: `https://${hermesBaseUrl}/api-v5/script/`,
       profileUrl: "/profile.html",
     };
     return {
       drupalDataset: undefined,
       exposeFunction,
-      hermesAPI: constructAPI(urls, false),
+      hermesAPI: constructAPI(urls),
       inDrupal,
       urls,
       user: {uid: "__everestui-standalone__"},